requirements.txt
django
djangorestframework
channels
daphne
together
python-dotenv
django-cors-headers
